requirements.txt
apify-client==1.7.0
requests==2.31.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.4
anthropic==0.28.0
pytrends==4.9.2
schedule==1.2.1
python-dotenv==1.0.0
